requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
DAWG-Python==0.7.2
docopt==0.6.2
fastapi==0.82.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
inexactsearch==1.0.2
intervaltree==3.1.0
ipymarkup==0.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
natasha==1.4.0
navec==0.10.0
numpy==1.21.6
pydantic==1.10.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyspellchecker==0.7.0
python-dotenv==0.21.0
PyYAML==6.0
razdel==0.5.0
silpa-common==0.3
slovnet==0.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.2.2
yargy==0.15.0
zipp==3.8.1
